Remove non-root user from docker file (#794)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Changes added in #790 to add a non-root user to final stage of docker
file introduced errors with PluginMonitor and file permissions that need
more in depth resolution. This temporary undoes the non-root user
changes to remove these errors. I future pull request is required to
update PluginMonitor to be able to run as non-root successfully. Those
changes will be in #793

Prior to this fix, docker image would fail to install plugins due to
permission issues. This was missed since the permission issues did not
exist in the dev container, which is built on a intermediate build
stage. Even after resolving the permission issues, the PluginMonitor
responsible for autostarting plugins when enabled, would not
successfully autostart the command plugin. #793 is planned to eventually
solve this problem
<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
NA
<!-- e.g. CAR-123 -->

## Motivation and Context
Remove bugs
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Local Integration Testing 
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.