Add build parameter to skip plugins (#796)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

These changes to V2X Hub allow for the user to select which plugins
should be built when running the build process for V2X Hub. The
CommandPlugin will always be built, as it is necessary for the web UI to
work. When running the initialization script, a prompt will appear in
the terminal, asking for which plugins to be selected. "All" selects all
the plugins (and is the default option), or typing out each plugin
selects those ones to be built.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key


[TELE-645](https://usdot-carma.atlassian.net/jira/software/c/projects/CS/boards/34?selectedIssue=TELE-645)

## Motivation and Context

As part of the effort for having the Telematics tool link to a corridor
of RSUs it is required that a lightweight version of V2X Hub be deployed
capable of streaming communicating with RSU’s that support NTCIP 1218
without the overhead of plugins which are required for communication
with the radio devices.

## How Has This Been Tested?

Local Deployment Testing

<img width="1788" height="848" alt="image"
src="https://github.com/user-attachments/assets/31d18b4b-7273-41d8-a579-3aa997b25f71"
/>


## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x ] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[TELE-645]:
https://usdot-carma.atlassian.net/browse/TELE-645?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: paulbourelly999 <paul.k.bourelly@leidos.com>