Fix UTC Time, update implementation for improved maintainability and Testability (#800)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This PR fixes the the issue described in VH-1485 :

Tried broadcasting TIM message, with accurate UTC time, and lead to no
message and no error. Most likely due to the message start time not
matching what V2x is expecting.

Checked logs in Komodo to find the time it is expecting for the current
time, and noticed it was incorrect.

This value was off by 300, which is 5 hours, which is the current hour
difference to UTC time.

This is most likely an issue in how the time is calculated on V2xhub.
Checking delta T of new year in this area/Reston, to time now in reston,
I got the V2x number. Checking delta T of UTC new years, to time now in
UTC, I got the TIM message number (which was calculated using datetime
in python)

 

Komodo: (444,461)

TIM message: (444,761)

Actual Time: (444,762)

**Solution:** The issue was caused by us of
[mktime](https://en.cppreference.com/w/cpp/chrono/c/mktime.html) which
converts a `tm` struct into a UTC `time_t` pointer. It assumes the `tm`
struct is in local time, but in our case the `tm` struct, which is
created from the TIM start and end time data is already in UTC. The
solution was to us
[timegm](https://man7.org/linux/man-pages/man3/timegm.3.html) which does
the same but assumes the input is UTC time already.

**Clean Up**: Additional to the functional fix, this PR includes updates
to move common non-stateful logic like evaluating whether a TIM is
currently active or reading an input TIM into free functions inside the
TimUtils.cpp to allow for improved maintainability and testability.
<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
[VH-1485
](https://usdot-carma.atlassian.net/browse/VH-1485)<!-- e.g. CAR-123 -->

## Motivation and Context
Fix the above described bug
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Unit testing 
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.