Create generate_local_trusted_certificates.sh (#805)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This PR updates the initialization script to ask the users whether to
generate local browser trusted certificates.
If **N** selected the self signed certificates for HTTPS access of the
web UI will be generated and stored in a shared volume for the **php**
and **v2xhub** services.
If **Y** selected the self signed certificates will be generated locally
using mkcert to add them to browser trusted certificates and stored in a
local directory. A message will be shown indicating to the user that
they should refer the the **ssl_datavolume** described in the volumes
section of the `docker-compose.yml` file to make the necessary
modifications to ensure **v2xhub** and **php** containers use these
locally certificates by mapping the **ssl_datavolume** to the local
directory where the locally generated certificates are stored.

Additional this PR includes PHP updates to redirect all HTTP(80) web ui
access to https(443).
<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
NA
<!-- e.g. CAR-123 -->

## Motivation and Context
Improve deployment process by allowing users to either
1) Run initialization script, configure local volume for local browser
trusted certificates
2) Generate certificates on container startup that are self signed

This PR also fixes a permissions issue with certificates generated for
DevContainer
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Running ./initialization locally
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.